% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomerData-class.R
\name{HomerData}
\alias{HomerData}
\title{Constructor for HomerData}
\usage{
HomerData(motif_data, logo_files, homer_dir)
}
\arguments{
\item{motif_data}{A data.frame of parsed HOMER results.}

\item{logo_files}{A character vector of paths to motif logo PNG files.}

\item{homer_dir}{The path to the HOMER output directory.}
}
\value{
A \code{HomerData} object.
}
\description{
Constructor for HomerData
}
