% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicviz-package.R
\docType{package}
\name{omicviz-package}
\alias{omicviz-package}
\alias{omicviz}
\title{omicviz: Visualization of HOMER Known Motif Discovery Results}
\description{
The omicviz package provides tools to parse and visualize HOMER known
motif discovery results. It creates publication-quality ComplexHeatmap
plots showing motif rank, transcription factor names, motif logos,
significance, and target enrichment.
}
\section{Main functions}{

\describe{
\item{\code{\link{readHomerDir}}}{Parse a HOMER output directory and
return a \code{HomerData} object.}
\item{\code{\link{vizHomerBar}}}{Create a ComplexHeatmap visualization
of known motif enrichment results.}
\item{\code{\link{vizVolcano}}}{Create a volcano plot from DESeq2
differential expression results.}
}
}

\section{Class}{

\describe{
\item{\code{\link{HomerData-class}}}{S4 class to store parsed HOMER
known motif results.}
}
}

\author{
\strong{Maintainer}: Author Huan Yang \email{hyang333@wisc.edu}

Authors:
\itemize{
  \item Author Huan Yang \email{hyang333@wisc.edu}
}

}
