% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HomerData-class.R
\name{show,HomerData-method}
\alias{show,HomerData-method}
\title{Show method for HomerData}
\usage{
\S4method{show}{HomerData}(object)
}
\arguments{
\item{object}{A HomerData object.}
}
\description{
Show method for HomerData
}
